Changes and Fixes to CLI, NH, ModelView.
Correction to situation of match: if ended, it is not available. Start adding the use of character and the show of situation from the cli: if client digits "show" or "character", he is able to view the islands, the schoolboards and the clouds situation or to use a character. -NOT COMPLETED-