Fixed tower choice & match end.
Fixed tower choice: if there are two players in the game, there must not be a grey tower color.
Fixed tower choice: now in cli are printed only the available (& not available) tower color(s).
Fixed match end: if the player ID winner received in the message is "-1", no one won the match because is still the waiting for start phase.