Lobby Join Fixed & CLI asssistant choice fixes.
*Lobby Join Fixed:*
After receiving a request of joining a match, it's checked if a lobby has reached the full capacity of players: for example can happen that a lobby with max 2 players (1 player in waiting) is shown to two clients. When one of the two joins it, if the other client try to join, is not accepted. It is because after the ReplyChosenLobby message the server check the availability.
To do that I:
- added a nack subobject to implement this reply;
- added cli print after the nack;
- moved in a method in NH the socket close;
- added the subobject case lobby_not_available in nack message of NH;
- added the check of the lobbies in the CH after receiving a chosen lobby message;
- added in CH the set of a lobby to end also when the match is started (it was missing in the OUT1 moment).

CLI:
- removed some fixed todo;
- fixed the assistant choice: when a player has in his deck only the same card as the cards already chosen in this round, he however can play them;
- added print of island when choosing mother nature island to move;
- added waiting method for the assistant choice phase;
- added waiting method for the cloud choice phase;
-